# back/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TECLAR — API",
    "version": "1.0",
    "description": "Gerada do mesmo contrato que o CONTRATO-API.md (js/nucleo/api.contrato.ts), por gerar-contrato.mjs. Rotas marcadas \"(ainda não implementada)\" ainda não existem no back.\n\nPara testar uma rota com cadeado: faça o POST /api/auth/login, copie o \"token\" da resposta e cole em **Authorize** (sem a palavra Bearer)."
  },
...
  "paths": {
...
    "/api/turmas/{id}/alunos": {
      "get": {
        "tags": [
          "Alunos (professor)"
        ],
        "operationId": "alunos.daTurma",
        "summary": "daTurma: Aluno[] — os alunos da turma, na ordem alfabética do nome",
        "description": "**Chamada no front:** `api.alunos.daTurma`\n\n**Resposta:** 200 Aluno[] — os alunos da turma, na ordem alfabética do nome\n\n**Identidade:**\n\nO token: a turma tem de ser da conta.\n\n**Regras de negócio no back (a tela não calcula):**\n\nAlunos com ClassID = turma e Ativo = TRUE; entrouEm = Alunos.Data_Cadastro, senhaDefinida = SenhaHash IS NOT NULL. Agregados prontos, null (nunca 0) para quem nunca treinou; a senha nunca sai.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Aluno[] — os alunos da turma, na ordem alfabética do nome",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "Exemplo 1",
                    "value": [
                      {
                        "id": "al-9f3k2",
                        "nome": "Ana Pires",
                        "entrouEm": "2026-02-01",
                        "senhaDefinida": true,
                        "totalSessoes": 12,
                        "wpmMedio": 39,
                        "precisaoMedia": 91,
                        "ultimaAtividade": "2026-10-03T12:00:00.000Z"
                      },
                      {
                        "id": "al-2m8qd",
                        "nome": "Davi Moreira",
                        "entrouEm": "2026-02-24",
                        "senhaDefinida": false,
                        "totalSessoes": 0,
                        "wpmMedio": null,
                        "precisaoMedia": null,
                        "ultimaAtividade": null
                      }
                    ]
                  }
                }
              }
            }
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — turma inexistente ou de outra conta",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/turmas/{id}/alunos/importar": {
      "post": {
        "tags": [
          "Alunos (professor)"
        ],
        "operationId": "alunos.importar",
        "summary": "importar: ResultadoImportacao — cada nome cai numa das três listas",
        "description": "**Chamada no front:** `api.alunos.importar`\n\n**Corpo:** { nomes } — a lista inteira numa requisição; a tela lê o CSV e manda só os nomes\n\n**Resposta:** 200 ResultadoImportacao — cada nome cai numa das três listas\n\n**Identidade:**\n\nO token: a turma tem de ser da conta.\n\n**Regras de negócio no back (a tela não calcula):**\n\nCada nome é aparado, com espaços internos reduzidos a um, e cai em falhas (vazio, item que não é texto, fora de 2 a 150 caracteres, ou \"Nome repetido na lista.\"), em jaEstavam (já na turma, sem distinguir maiúscula e acento) ou vira linha nova de Alunos com SenhaHash NULL. Importação parcial é permitida, e a rota pode ser chamada de novo com a lista completa.\n\n**Notas:**\n\nÉ a única entrada do aluno na turma. Aluno removido volta como linha nova, sem senha nem histórico.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "{ nomes } — a lista inteira numa requisição; a tela lê o CSV e manda só os nomes",
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              },
              "examples": {
                "exemplo1": {
                  "summary": "Exemplo 1",
                  "value": {
                    "nomes": [
                      "Ana Pires",
                      "Bruno Sato",
                      "ana  pires",
                      "",
                      "Carla Nunes"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "ResultadoImportacao — cada nome cai numa das três listas",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "Exemplo 1",
                    "value": {
                      "adicionados": [
                        {
                          "id": "al-7c1pz",
                          "nome": "Bruno Sato",
                          "entrouEm": "2026-10-05",
                          "senhaDefinida": false,
                          "totalSessoes": 0,
                          "wpmMedio": null,
                          "precisaoMedia": null,
                          "ultimaAtividade": null
                        }
                      ],
                      "jaEstavam": [
                        "Ana Pires"
                      ],
                      "falhas": [
                        {
                          "nome": "ana  pires",
                          "motivo": "Nome repetido na lista."
                        },
                        {
                          "nome": "",
                          "motivo": "Nome vazio."
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "`DADOS_INVALIDOS` — `nomes` não é uma lista, tem mais de 500 itens, ou o corpo traz professorId",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "DADOS_INVALIDOS",
                    "value": {
                      "mensagem": "...",
                      "codigo": "DADOS_INVALIDOS"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — turma inexistente ou de outra conta",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/turmas/{id}/alunos/{alunoId}": {
      "delete": {
        "tags": [
          "Alunos (professor)"
        ],
        "operationId": "alunos.remover",
        "summary": "remover: (sem corpo)",
        "description": "**Chamada no front:** `api.alunos.remover`\n\n**Resposta:** 204 (sem corpo)\n\n**Identidade:**\n\nO token: a turma tem de ser da conta (senão 404). O :alunoId é o aluno-ALVO — sobre quem a ação é —, e não quem pede.\n\n**Regras de negócio no back (a tela não calcula):**\n\nApaga a linha de Alunos; as sessões dele vão junto (SessionsProf tem ON DELETE CASCADE). Ele deixa de conseguir entrar na hora.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alunoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "(sem corpo)"
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — turma de outra conta, ou o aluno não é desta turma",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/turmas/{id}/alunos/{alunoId}/zerar-senha": {
      "post": {
        "tags": [
          "Alunos (professor)"
        ],
        "operationId": "alunos.zerarSenha",
        "summary": "zerarSenha: (sem corpo)",
        "description": "**Chamada no front:** `api.alunos.zerarSenha`\n\n**Resposta:** 204 (sem corpo)\n\n**Identidade:**\n\nO token: a turma tem de ser da conta (senão 404). O :alunoId é o aluno-ALVO, como no DELETE.\n\n**Regras de negócio no back (a tela não calcula):**\n\nGrava Alunos.SenhaHash = NULL: o próximo login vira primeiro acesso, e o histórico fica. Zerar quem já está com SenhaHash NULL responde 204 igual.\n\n**Notas:**\n\nAté o próximo login, quem tiver o código e o nome cria a senha; a tela avisa disso.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alunoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "(sem corpo)"
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — turma de outra conta, ou o aluno não é desta turma",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/turmas/{id}/alunos/{alunoId}/desempenho": {
      "get": {
        "tags": [
          "Alunos (professor)"
        ],
        "operationId": "alunos.desempenho",
        "summary": "desempenho: DesempenhoAluno — os agregados e as sessões do aluno",
        "description": "**Chamada no front:** `api.alunos.desempenho`\n\n**Resposta:** 200 DesempenhoAluno — os agregados e as sessões do aluno\n\n**Identidade:**\n\nO token: a turma tem de ser da conta (senão 404). O :alunoId é o aluno-ALVO, como no DELETE.\n\n**Regras de negócio no back (a tela não calcula):**\n\nAs sessões do aluno nesta turma e os agregados sobre elas.\n\n**Notas:**\n\nNenhuma tela usa (o modal de relatórios usa GET /turmas/:id/alunos/:alunoId/sessoes).",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "alunoId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "DesempenhoAluno — os agregados e as sessões do aluno",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "Exemplo 1",
                    "value": {
                      "alunoId": "al-9f3k2",
                      "totalSessoes": 2,
                      "wpmMedio": 38,
                      "precisaoMedia": 94,
                      "sessoes": [
                        {
                          "id": "ses-1",
                          "exerciseId": "ex-prof-1",
                          "alunoId": "al-9f3k2",
                          "wpm": 40,
                          "precisao": 95,
                          "tempoSegundos": 58,
                          "acertos": 76,
                          "erros": 4,
                          "concluida": true,
                          "data": "2026-02-18T14:10:00.000Z"
                        }
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "`TOKEN_INVALIDO` — token ausente, vencido ou inválido",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "TOKEN_INVALIDO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "TOKEN_INVALIDO"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "`NAO_ENCONTRADO` — turma de outra conta, ou o aluno não é desta turma",
            "content": {
              "application/json": {
                "examples": {
                  "exemplo1": {
                    "summary": "NAO_ENCONTRADO",
                    "value": {
                      "mensagem": "...",
                      "codigo": "NAO_ENCONTRADO"
                    }
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}